// src/app/shared/enums/PaymentType.ts
export enum PaymentType {
  // Loại kỳ thanh toán
  ALL = 'ALL',
  PART = 'PART',
  SAME_TYPE = 'SAME_TYPE',
  ANY = 'ANY',
  ALL_OR_PART = 'ALL_OR_PART',
  ONE = 'ONE',
  MANY = 'MANY'
}

export enum PaymentPeroidType {
  NEAR_TO_FAR = 'NEAR_TO_FAR',
  FAR_TO_NEAR = 'FAR_TO_NEAR',
  ANY = 'ANY',
  MULTIPE = 'MULTIPE'
}

export enum PaymentPriority {
  ELECTRIC = 'ELECTRIC',
  ANY = 'ANY'
}